CommentLike 스키마 모델 생성

+++ src/routers/post.router.js
@@ -13,7 +13,6 @@ postsRouter.post(
   requireAccessToken,
   postCreateValidator,
   async (req, res, next) => {
-    //console.log(req)
     try {
       const { userId } = req.user;
       const { title, content, regionId, imageUrl } = req.body;
@@ -54,7 +53,6 @@ postsRouter.get('/posts', requireAccessToken, async (req, res, next) => {
     }
 
     let data = await prisma.post.findMany({
-      where: { userId },
       orderBy: {
         createdAt: sort,
       },
@@ -62,7 +60,6 @@ postsRouter.get('/posts', requireAccessToken, async (req, res, next) => {
         user: true,
       },
     });
-
     data = data.map((post) => {
       return {
         postId: post.postId,
@@ -74,7 +71,6 @@ postsRouter.get('/posts', requireAccessToken, async (req, res, next) => {
         updatedAt: post.updatedAt,
       };
     });
-
     return res
       .status(HTTP_STATUS.OK)
       .json({ status: HTTP_STATUS.OK, message: POST_MESSAGES.POST_LIST, data });
@@ -92,31 +88,65 @@ postsRouter.get('/:id', async (req, res, next) => {
 
     let data = await prisma.post.findUnique({
       where: { postId: Number(postId) },
-      //include: { user: true},
+      include: {
+        user: {
+          select: {
+            nickname: true
+          }
+        },
+        Comment: {
+          take: 3,
+          include: {
+            user: {
+              select: {
+                nickname: true
+              }
+            }
+          }
+        }
+      }
     });
-
+    
     if (!data) {
-      return res.status(HTTP_STATUS.NOT_FOUND).json({
-        status: HTTP_STATUS.NOT_FOUND,
-        message: POST_MESSAGES.POST_NOT_FOUND,
-        data,
-      });
+      return res
+        .status(HTTP_STATUS.NOT_FOUND)
+        .json({
+          status: HTTP_STATUS.NOT_FOUND,
+          message: POST_MESSAGES.POST_NOT_FOUND,
+          data,
+        });
     }
+
+    let comments = data.Comment.map(comment => {
+      return {
+        commentId: comment.commentId,
+        userId: comment.userId,
+        postId: comment.postId,
+        comment: comment.comment,
+        createdAt: comment.createdAt,
+        updatedAt: comment.updatedAt,
+        nickname: comment.user.nickname // 댓글 작성자의 닉네임을 새로운 필드로 추가합니다.
+      };
+    });
 
     data = {
       postId: data.postId,
       title: data.title,
       content: data.content,
+      nickname: data.user.nickname,
       regionId: data.regionId,
       imageUrl: data.imageUrl,
       createdAt: data.createdAt,
       updatedAt: data.updatedAt,
+      comment: comments
     };
-    return res.status(HTTP_STATUS.OK).json({
-      status: HTTP_STATUS.OK,
-      message: POST_MESSAGES.POST_DETAIL,
-      data,
-    });
+    return res
+      .status(HTTP_STATUS.OK)
+      .json({
+        status: HTTP_STATUS.OK,
+        message: POST_MESSAGES.POST_DETAIL,
+        data,
+      });
   } catch (error) {
     next(error);
   }
@@ -231,6 +261,7 @@ postsRouter.patch(
       const result = await prisma.post.findUnique({
         where: { postId: +postId },
       });
+
       if (!result) {
         return res.status(400).json({ message: '게시글이 존재하지 않습니다.' });
       }
@@ -295,18 +326,19 @@ postsRouter.patch(
       }
 
       // 좋아요 존재 여부 확인
-      const likeData = await prisma.like.findFirst({
+      const likeData = await prisma.commentLike.findFirst({
         where: {
           commentId: +commentId,
+          postId :+postId,
           userId,
         },
       });
 
       if (likeData) {
         // 좋아요 삭제
-        await prisma.like.delete({
+        await prisma.commentLike.delete({
           where: {
-            likeId: likeData.likeId,
+            commentlikeId: likeData.commentlikeId,
           },
         });
         return res
@@ -314,7 +346,7 @@ postsRouter.patch(
           .json({ message: '댓글 좋아요가 삭제되었습니다.' });
       } else {
         // 좋아요 생성
-        await prisma.like.create({
+        await prisma.commentLike.create({
           data: {
             commentId: +commentId,
             postId: +postId,
@@ -358,10 +390,22 @@ postsRouter.post(
           comment: comment,
         },
       });
-
+      const data = await prisma.user.findUnique({
+        where : {userId: userId}
+      })
+      const result = {
+        comment: commentData.comment,
+        commentId: commentData.commentId,
+        createdAt: commentData.createdAt,
+        updatedAt: commentData.updatedAt,
+        postId: commentData.postId,
+        userId: commentData.userId,
+        nickname: data.nickname
+      }
+      console.log(result)
       return res
         .status(200)
-        .json({ message: '댓글 생성이 완료했습니다.', data: commentData });
+        .json({ message: '댓글 생성이 완료했습니다.', data: result });
     } catch (error) {
       next(error);
     }
@@ -424,18 +468,16 @@ postsRouter.patch(
       const { postId, commentId } = req.params;
       const { userId } = req.user;
       const { comment } = req.body;
-
       const existingComment = await prisma.comment.findFirst({
         where: {
           postId: +postId,
-          contentId: +commentId,
+          commentId: +commentId,
         },
       });
 
       if (!existingComment) {
         return res.status(404).json({ message: '댓글을 찾을 수 없습니다.' });
       }
-
       if (existingComment.userId !== userId) {
         return res
           .status(403)
@@ -444,7 +486,7 @@ postsRouter.patch(
 
       const updatedComment = await prisma.comment.update({
         where: {
-          contentId: +commentId,
+          commentId: +commentId,
           postId: +postId,
           userId: +userId,
         },
@@ -452,10 +494,12 @@ postsRouter.patch(
           comment: comment,
         },
       });
-      return res.status(200).json({
-        message: '댓글이 성공적으로 수정되었습니다.',
-        data: updatedComment,
-      });
+      return res
+        .status(200)
+        .json({
+          message: '댓글이 성공적으로 수정되었습니다.',
+          data: updatedComment,
+        });
     } catch (error) {
       next(error);
     }
@@ -472,9 +516,8 @@ postsRouter.delete(
 
       const data = await prisma.comment.findFirst({
         where: {
-          contentId: +commentId,
-          postId: +postId,
-          userId: userId,
+          commentId: +commentId,
+          postId: +postId        
         },
       });
 
@@ -482,16 +525,24 @@ postsRouter.delete(
         return res.status(400).json({ message: '댓글이 없습니다.' });
       }
 
+      if (data.userId !== userId) {
+        return res
+          .status(403)
+          .json({ message: '댓글을 삭제할 수 있는 권한이 없습니다.' });
+      }
+
       const deleteComment = await prisma.comment.delete({
         where: {
-          contentId: +commentId,
+          commentId: +commentId,
           userId: userId,
         },
       });
-      return res.status(200).json({
-        message: '댓글이 성공적으로 삭제되었습니다.',
-        data: deleteComment,
-      });
+      return res
+        .status(200)
+        .json({
+          message: '댓글이 성공적으로 삭제되었습니다.',
+          data: deleteComment,
+        });
     } catch (error) {
       next(error);
     }
